pyproject.toml
[tool.poetry]
name = "notionpub"
version = "0.1.0"
description = "publish files to notion"
authors = ["Tyler M. Kontra <tyler@tylerkontra.com>"]
license = "GPL"
readme = "README.md"

packages = [
    { include = "notionpub", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
md2notion = { path = "src/vendor/md2notion/", develop = false }
marshmallow = "^3.20.1"
pyyaml = "^6.0.1"
pydantic = "^2.5.3"
notion-client = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
